feat: Allow opening a bookmarked location in $EDITOR
Summary:
Previously, the default action when you hit ENTER was to cd into the directory
of the selected bookmark. This could be improved in 2 ways:
1. Adding an option to open the bookmarked location in $EDITOR makes more sense
when the bookmark points to a file; also fine with directories as it provides a
quick way to "jump into the project".
2. Changing the default behaviour of ENTER to "open a file in $EDITOR if it's
set otherwise cd to the directory" ensures that the behavior is reasonable even
when $EDITOR is not set.